% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_aoi.R
\name{center_times}
\alias{center_times}
\title{Normalize time by the onset}
\usage{
center_times(df)
}
\arguments{
\item{df}{df that has subject_id, dataset_id, trial_id and times}
}
\value{
df_out df that has the normalized times
}
\description{
Normalize time by the onset
}
