% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{create_emtpy_table}
\alias{create_emtpy_table}
\title{Create empty table}
\usage{
create_emtpy_table(table_type)
}
\arguments{
\item{table_type}{character TODO}
}
\value{
TODO
}
\description{
Create empty table
}
