% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{generate_aoi}
\alias{generate_aoi}
\title{Generate aoi data from xy data}
\usage{
generate_aoi(dir)
}
\arguments{
\item{dir}{Directory with xy data and metadata}
}
\description{
Generate aoi data from xy data
}
