% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{get_json_colnames}
\alias{get_json_colnames}
\title{Fetching the list of column names in each table according to the json file}
\usage{
get_json_colnames(table_type)
}
\arguments{
\item{table_type}{the type of table, can be one of this six types:
xy_data, aoi_data, participants, trials, dataset, aoi_regions}
}
\value{
colnames_json -- the list of column names
}
\description{
Fetching the list of column names in each table according to the json file
}
\examples{
\dontrun{
colnames_json <- get_json_colnames(table_type = "aoi_data")
}

}
