% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{get_peekjson}
\alias{get_peekjson}
\title{get json file from peekbank github}
\usage{
get_peekjson()
}
\value{
peekjson -- the organized dataframe from json file
}
\description{
get json file from peekbank github
}
\examples{
\dontrun{
peekjson <- get_peekjson()
}

}
