% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{list_coding_methods}
\alias{list_coding_methods}
\title{Get the coding method list from json schema file}
\usage{
list_coding_methods()
}
\value{
a list of strings indicating allowed coding methods
}
\description{
Get the coding method list from json schema file
}
\examples{
\dontrun{
coding_methods <- list_coding_methods()
}
}
