% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{list_language_choices}
\alias{list_language_choices}
\title{List current allowed language choices for db import}
\usage{
list_language_choices()
}
\value{
language_list TODO
}
\description{
List current allowed language choices for db import
}
\examples{
\dontrun{
language_list <- list_language_choices()
}

}
