% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{map_columns}
\alias{map_columns}
\title{Function for map raw data columns to processed table columns}
\usage{
map_columns(raw_data, raw_format, table_type)
}
\arguments{
\item{raw_data}{raw dara frame}

\item{raw_format}{source of the eye-tracking data, e.g. "tobii"}

\item{table_type}{type of processed table, e.g. "xy_data" | "aoi_table"}
}
\value{
processed data frame with specified column names
}
\description{
Function for map raw data columns to processed table columns
}
\examples{
\dontrun{
df_xy_data <- map_columns(raw_data = raw_data, raw_format = "tobii", table_type = "xy_data")
df_aoi_data <- map_columns(raw_data = raw_data, raw_format = "tobii", table_type = "aoi_data")
}

}
