% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{process_smi}
\alias{process_smi}
\title{Process smi raw data}
\usage{
process_smi(dir, file_ext = ".txt")
}
\arguments{
\item{dir}{Directory with raw data}

\item{file_ext}{TODO}
}
\description{
Process smi raw data
}
