% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{process_smi_file}
\alias{process_smi_file}
\title{Process an individual smi raw data file}
\usage{
process_smi_file(x, dir, stims_to_remove_chars = c(".avi"),
  stims_to_keep_chars = c("_"), possible_delims = c("\\t", ","),
  stimulus_coding = "stim_column")
}
\arguments{
\item{stimulus_coding}{}
}
\description{
Process an individual smi raw data file
}
