% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{process_tobii}
\alias{process_tobii}
\title{Process tobii raw data
dataset specific: datasets}
\usage{
process_tobii(
  dir_raw,
  dataset_name = "sample_data",
  dataset_type = "automated"
)
}
\arguments{
\item{dir_raw}{Directory with raw data}
}
\description{
Process tobii raw data
dataset specific: datasets
}
