% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{put_processed_data}
\alias{put_processed_data}
\title{Put processed data for specific peekbank dataset on OSF}
\usage{
put_processed_data(token, lab_dataset_id, path = ".", osf_address = "pr6wu")
}
\arguments{
\item{lab_dataset_id}{Specific ID occurring in the file hierarchy of the relevant OSF repo.}

\item{path}{Where the data live on your own machine.}

\item{osf_address}{pr6wu for peekbank.}
}
\description{
Put processed data for specific peekbank dataset on OSF
}
