% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_aoi.R
\name{resample_times}
\alias{resample_times}
\title{Resample times to be consistent across labs}
\usage{
resample_times(df)
}
\arguments{
\item{dir}{df that has subject_id, dataset_id, trial_id and times}
}
\value{

}
\description{
Resample times to be consistent across labs
}
