% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_aoi.R
\name{round_times}
\alias{round_times}
\title{Round times to our specified sample rate to be consistent across labs}
\usage{
round_times(df)
}
\arguments{
\item{df}{df that has subject_id, dataset_id, trial_id and times}
}
\value{
TODO
}
\description{
Round times to our specified sample rate to be consistent across labs
}
