% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{save_table}
\alias{save_table}
\title{supporting function for saving processed table as csv files in
processed_data/}
\usage{
save_table(df_procd, table_type)
}
\arguments{
\item{df_procd}{processed data frame}

\item{table_type}{type of table to be saved}
}
\description{
supporting function for saving processed table as csv files in
processed_data/
}
