% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_trial_uniqueness_constraint}
\alias{validate_trial_uniqueness_constraint}
\title{Check if within aoi_timepoints table, there is no duplication in all the administration_ids
  associated with each individual trial_id}
\usage{
validate_trial_uniqueness_constraint(df_aoi_timepoints)
}
\arguments{
\item{df_table}{the aoi_timepoints dataframe}
}
\value{
an empty string when all the administration_ids are unique within each trial_id;
  Otherwise, the error message will be returned.
}
\description{
Check if within aoi_timepoints table, there is no duplication in all the administration_ids
  associated with each individual trial_id
}
\examples{
\dontrun{
is_valid <- validate_table(df_table = df_table, table_type = "xy_data")
}

}
