% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{visualize_for_db_import}
\alias{visualize_for_db_import}
\title{Generate time course plots after all the processed tables are validated}
\usage{
visualize_for_db_import(dir_csv, lab_dataset_id, is_save = FALSE)
}
\arguments{
\item{dir_csv}{the folder directory containing all the csv files,
the path should end in "processed_data"}

\item{lab_dataset_id}{TODO}

\item{is_save}{TODO}
}
\value{
TODO
}
\description{
Generate time course plots after all the processed tables are validated
}
\examples{
\dontrun{
visualize_for_db_import(dir_csv = "./processed_data")
}

}
