% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy_trim.R
\name{xy_trim}
\alias{xy_trim}
\title{This function trims the xy_data such that only gaze coordinates that fall within the stimuli space is preserved
To make this work, a line of code will need to be added to each import script after center_time_on_pod():
%>% xy_trim(datasets)}
\usage{
xy_trim(xy_data, datasets)
}
\arguments{
\item{xy_data}{the pre-trim xy data generated from the import script}

\item{datasets}{the peekDS formatted datasets.csv that contains screen dimension info}
}
\description{
This function trims the xy_data such that only gaze coordinates that fall within the stimuli space is preserved
To make this work, a line of code will need to be added to each import script after center_time_on_pod():
%>% xy_trim(datasets)
}
