% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{get_json_fields}
\alias{get_json_fields}
\title{Fetching the list of field names and requirements in each table according to the peekds json file}
\usage{
get_json_fields(table_type)
}
\arguments{
\item{table_type}{the type of table, can be one of the following types:
aoi_timepoints, aoi_region_sets, administrations, subjects, trials, datasets, stimuli}
}
\value{
fieldnames -- the list of field names
}
\description{
Fetching the list of field names and requirements in each table according to the peekds json file
}
\examples{
\dontrun{
fields_json <- get_json_fields(table_type = "aoi_timepoints")
}

}
