% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{get_peekjson}
\alias{get_peekjson}
\title{parse json file from peekbank github into a dataframe}
\usage{
get_peekjson()
}
\value{
the organized dataframe from schema json file
}
\description{
parse json file from peekbank github into a dataframe
}
\examples{
\dontrun{
peekjson <- get_peekjson()
}

}
