% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{is_table_required}
\alias{is_table_required}
\title{Is this table required?}
\usage{
is_table_required(table_type, coding_method)
}
\arguments{
\item{table_type}{character TODO}

\item{coding_method}{character TODO}
}
\value{
TODO
}
\description{
Is this table required?
}
\examples{
\dontrun{
is_required <- is_table_required(table_type = "xy_timepoints",
                                 coding_method = "manual gaze coding")
}

}
