% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{is_table_required}
\alias{is_table_required}
\title{Is this table required?}
\usage{
is_table_required(table_type, dataset_type = "auto")
}
\arguments{
\item{table_type}{character}

\item{dataset_type}{character}
}
\value{

}
\description{
Is this table required?
}
