% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{list_ds_tables}
\alias{list_ds_tables}
\title{List the tables required for different datasets}
\usage{
list_ds_tables(dataset_type = "automated")
}
\arguments{
\item{dataset_type}{}
}
\description{
List the tables required for different datasets
}
