% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{list_ds_tables}
\alias{list_ds_tables}
\title{List the tables required for datasets with different coding method}
\usage{
list_ds_tables(coding_method = "eyetracking")
}
\arguments{
\item{coding_method}{character TODO}
}
\value{
TODO
}
\description{
List the tables required for datasets with different coding method
}
\examples{
\dontrun{
table_list <- list_ds_tables(coding_method = "manual gaze coding")
}

}
