% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{process_to_xy}
\alias{process_to_xy}
\title{Process directory of raw data to xy data}
\usage{
process_to_xy(format, dir)
}
\arguments{
\item{format}{One of "tobii", "smi", "eyelink"}

\item{dir}{Directory with raw data}
}
\value{
TODO
}
\description{
Process directory of raw data to xy data
}
